Add wildcard handling for ZMTX.
The user can specify a list of files or wildcard patterns
on the command line, and ZMTX will use CP/M directory functions
to find matching files.

To avoid dynamic memory allocation, the filename buffer
is statically allocated and is limited to MAX_MATCHES,
currently 512 entries. (Can be changed in fileio.h.)